% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_zonal_statistics.R
\name{compute_zonal_tasks}
\alias{compute_zonal_tasks}
\title{Compute zonal tasks}
\usage{
compute_zonal_tasks(zonal_tasks, g_var, db_file)
}
\arguments{
\item{zonal_tasks}{A tibble of zonal tasks created by \code{create_zonal_tasks}}

\item{g_var}{Variable name (string) that unique identifies each feature at the sf object.}

\item{db_file}{A path and file name for the SQLite database.}
}
\value{
An SQLite database.
}
\description{
Compute zonal tasks
}
