% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_zonal_tasks.R
\name{create_zonal_tasks}
\alias{create_zonal_tasks}
\title{Create tasks of zonal statistics}
\usage{
create_zonal_tasks(
  nc_files_list,
  nc_chunk_size,
  sf_geom,
  sf_chunck_size,
  zonal_functions
)
}
\arguments{
\item{nc_files_list}{A vector of NetCDF files addresses.}

\item{nc_chunk_size}{Chunk size for raster layers processing.}

\item{sf_geom}{An sf object.}

\item{sf_chunck_size}{Chunk size for sf features processing.}

\item{zonal_functions}{A vector of zonal statistics to be computed.}
}
\value{
A tibble where each row is a zonal statistic that should be computed using a chunk of raster layers and a chunk of sf features.
}
\description{
Create tasks of zonal statistics
}
