% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadm41_moz.R
\docType{data}
\name{gadm41_moz}
\alias{gadm41_moz}
\title{Mozambique administrative boundaries, level 3}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 413 rows and 3 columns.
}
\usage{
gadm41_moz
}
\description{
A sf object.
}
\keyword{datasets}
