% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_task.R
\name{compute_task}
\alias{compute_task}
\title{Compute a zonal task}
\usage{
compute_task(rst, pol, pop = NULL, g_var, fn_name, db_file)
}
\arguments{
\item{rst}{The rst variable from a zonal tasks tibble}

\item{pol}{The pol variable from a zonal tasks tibble}

\item{pop}{Optional. Population weighting raster}

\item{g_var}{Variable name (string) that unique identifies each feature at the sf object.}

\item{fn_name}{The fn_name variable from a zonal tasks tibble}

\item{db_file}{Path and file name to SQLite database.}
}
\value{
Writes the results of the zonal taks into the SQLite database and return a TRUE value.
}
\description{
Compute a zonal task
}
